pyproject.toml
[project]
name = "stat-xplore-mcp"
version = "0.1.0"
description = "MCP server and API for DWP Stat-Xplore"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.28.0",
    "mcp>=1.0.0",
    "modal>=0.68.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.1",
    "rich>=13.9.0",
    "uvicorn>=0.34.0",
]

[project.scripts]
stat-xplore-mcp = "stat_xplore_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stat_xplore_mcp"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
